requirements.txt
click==8.0.1
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
psycopg2-binary==2.9.1
pymongo==3.12.0
python-decouple==3.4
SQLAlchemy==1.4.22
Werkzeug==2.0.1
WTForms==2.3.3
